backend/requirements.txt
# FastAPI and server
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# Existing dependencies from original project
torch>=2.0.0
transformers>=4.36.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
